+++ test/simulations/logRebalance.js
@@ -1,6 +1,6 @@
 const { decodeHexString, saveToCSV } = require("./common");
 
 const packedHexString = process.argv[2];
-const args = decodeHexString(packedHexString, ["int24", "uint256", "uint256"]);
+const args = decodeHexString(packedHexString, ["int256", "uint256", "uint256"]);
 
 saveToCSV("rebalances", `${args.join(",")}\n`);
